+++ src/Package/Constants.cs
@@ -1,13 +1,13 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ExperimentalFeatures
 {
     public class Constants
     {
-        public const string LiveFeedPath = "";
+        public const string LiveFeedPath = "https://raw.githubusercontent.com/ligershark/ExperimentalFeatures/master/extensions.json";
+        public static readonly string LiveFeedCachePath = Environment.ExpandEnvironmentVariables("%localAppData%\\" + Vsix.Name + "\\feed.json");
+        public static readonly string LogFile = Environment.ExpandEnvironmentVariables("%localAppData%\\" + Vsix.Name + "\\installer.log");
+
+        public const double UpdateInterval = 3;
     }
 }